All kinds of crud.. really should commit more since I don't remember what I did...
But one thing is the reorganizing of the model for groups so that the form doesn't crash.

However, the form element is filling with the groups now even if you don't have an inventory type.

Will need to fix this, probably just not use that form element in the template.