Fixes #398
When creating an Issue, an ajax call would be applied for the 'Name' field for auto-completing Machine names. After some tweaking,
the load time for the auto complete is cut roughly in half. Created new methods and calls to help optimize the load time.
Left in the previous methods and calls, but commented them out.